+++ CHANGELOG.md
@@ -6,16 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Pushdown Automata & Regular Expression modules are now available.
 - Language is saved across sessions.
 - Regular expression scope highlighting.
 - Cross-window communication through BroadcastChannel.
 - Color theme preset loading.
 - MIT and Duke theme presets.
-- Added linting tests.
-- Added Travic CI build pipeline.
+- Added linting tests to enforce code style.
+- Added Travic CI build pipeline for more build automation.
 
 ### Changed
 - Have actual changelog that is not code :P
+- Experimental mode is now LIVE! :D
 
 ### Removed
 - Old app mode will be replaced by experimental mode permanently :(